refactor: remove entity framework from building blocks
EF is good for write side, but it's not suit for rich domain modeling. after this commit, I will start to use MongoDB for read and write side. Maybe in the future I use event store for write side.  But, for now, it adds complexity of versioning to the app and make it too hard to manage as an open source project